New titlebar
Added a new titlebar. Working on improving the AddDialogue dialogue, adding a feature so that little dots appear for each item under their day on the calendar, adding a dropdown menu or whatever it is called in the top left corner - it will lead to the Weather section of the app and plan on adding connection to an SQLite database to store the listView items or just store them in a .json if all hope gets lost.